cmudb/behavior_modeling/requirements.txt
# Python dependencies required for development
numpy
psutil
pyyaml
setuptools
pandas
sklearn>=1.0.0
pydotplus
lightgbm
isort
black
flake8
dataclasses
bandit
pylint
mypy
